# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  models.CreateScanRequest:
    properties:
      content:
        description: 扫描内容
        type: string
      imageId:
        description: 关联的图片ID
        type: string
      success:
        description: 扫描是否成功
        type: boolean
    required:
    - content
    - success
    type: object
...
  models.UpdateScanRequest:
    properties:
      content:
        description: 扫描内容
        type: string
      imageId:
        description: 关联的图片ID
        type: string
      success:
        description: 扫描是否成功
        type: boolean
    required:
    - content
    - success
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: 这是一个基于Go和Gin框架的后端API服务
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Go Backend API
  version: "1.0"
paths:
...
  /scans:
    get:
      consumes:
      - application/json
      description: 获取系统中所有扫描记录的列表
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              count:
                type: integer
              data:
                items:
                  type: object
                type: array
              success:
                type: boolean
            type: object
        "500":
          description: Internal Server Error
          schema:
            properties:
              message:
                type: string
              success:
                type: boolean
            type: object
      summary: 获取所有扫描记录
      tags:
      - scans
    post:
      consumes:
      - application/json
      description: 创建新的扫描记录
      parameters:
      - description: 扫描记录信息
        in: body
        name: scan
        required: true
        schema:
          $ref: '#/definitions/models.CreateScanRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            properties:
              data:
                type: object
              success:
                type: boolean
            type: object
        "400":
          description: Bad Request
          schema:
            properties:
              message:
                type: string
              success:
                type: boolean
            type: object
        "500":
          description: Internal Server Error
          schema:
            properties:
              message:
                type: string
              success:
                type: boolean
            type: object
      summary: 创建扫描记录
      tags:
      - scans
  /scans/{id}:
    get:
      consumes:
      - application/json
      description: 根据扫描记录ID获取扫描记录详细信息
      parameters:
      - description: 扫描记录ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              data:
                type: object
              success:
                type: boolean
            type: object
        "400":
          description: Bad Request
          schema:
            properties:
              message:
                type: string
              success:
                type: boolean
            type: object
        "404":
          description: Not Found
          schema:
            properties:
              message:
                type: string
              success:
                type: boolean
            type: object
        "500":
          description: Internal Server Error
          schema:
            properties:
              message:
                type: string
              success:
                type: boolean
            type: object
      summary: 根据ID获取扫描记录
      tags:
      - scans
    put:
      consumes:
      - application/json
      description: 根据ID更新扫描记录信息
      parameters:
      - description: 扫描记录ID
        in: path
        name: id
        required: true
        type: integer
      - description: 扫描记录信息
        in: body
        name: scan
        required: true
        schema:
          $ref: '#/definitions/models.UpdateScanRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              data:
                type: object
              success:
                type: boolean
            type: object
        "400":
          description: Bad Request
          schema:
            properties:
              message:
                type: string
              success:
                type: boolean
            type: object
        "404":
          description: Not Found
          schema:
            properties:
              message:
                type: string
              success:
                type: boolean
            type: object
        "500":
          description: Internal Server Error
          schema:
            properties:
              message:
                type: string
              success:
                type: boolean
            type: object
      summary: 更新扫描记录
      tags:
      - scans
  /scans/export:
    get:
      consumes:
      - application/json
      description: 将扫描记录导出为Excel文件
      parameters:
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 10000
        description: 每页数量
        in: query
        name: page_size
        type: integer
      - default: desc
        description: 排序方式
        in: query
        name: order
        type: string
      - default: create_time
        description: 排序字段
        in: query
        name: order_by
        type: string
      produces:
      - application/vnd.openxmlformats-officedocument.spreadsheetml.sheet
      responses:
        "200":
          description: Excel文件
          schema:
            type: file
        "400":
          description: Bad Request
          schema:
            properties:
              message:
                type: string
              success:
                type: boolean
            type: object
        "500":
          description: Internal Server Error
          schema:
            properties:
              message:
                type: string
              success:
                type: boolean
            type: object
      summary: 导出扫描记录为Excel
      tags:
      - scans
  /scans/page:
    get:
      consumes:
      - application/json
      description: 根据分页参数获取扫描记录列表
      parameters:
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 10
        description: 每页数量
        in: query
        name: page_size
        type: integer
      - default: desc
        description: 排序方式
        in: query
        name: order
        type: string
      - default: create_time
        description: 排序字段
        in: query
        name: order_by
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              data:
                items:
                  type: object
                type: array
              pagination:
                type: object
              success:
                type: boolean
            type: object
        "400":
          description: Bad Request
          schema:
            properties:
              message:
                type: string
              success:
                type: boolean
            type: object
        "500":
          description: Internal Server Error
          schema:
            properties:
              message:
                type: string
              success:
                type: boolean
            type: object
      summary: 分页获取扫描记录列表
      tags:
      - scans
...
securityDefinitions:
  BasicAuth:
    type: basic
swagger: "2.0"